server/src/test/java/com/exadel/controllers/ContractControllerTest.java
package com.exadel.controllers;

import com.exadel.dto.ContractDto;
import com.exadel.services.ContractService;
import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Mockito;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.mock.web.MockHttpServletResponse;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.MvcResult;
import org.springframework.test.web.servlet.RequestBuilder;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;

import java.util.ArrayList;
import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.Assert.assertEquals;
import static org.mockito.Mockito.when;

@WebMvcTest(ContractController.class)
@ExtendWith(MockitoExtension.class)
public class ContractControllerTest {
    private static final String TEST_URL = "/contracts";

    @Autowired
    MockMvc mockMvc;

    @MockBean
    ContractService contractService;

    @Test
    public void testAddContract() throws Exception {
        ContractDto contractDtoMock = ContractDto.builder().build();

        String inputJson = this.toJson(contractDtoMock);
        when(contractService.save(Mockito.any(ContractDto.class))).thenReturn(contractDtoMock);

        RequestBuilder requestBuilder = MockMvcRequestBuilders
                .post(TEST_URL)
                .accept(MediaType.APPLICATION_JSON).content(inputJson)
                .contentType(MediaType.APPLICATION_JSON);

        MvcResult result = mockMvc.perform(requestBuilder).andReturn();
        MockHttpServletResponse response = result.getResponse();
        String outputJson = response.getContentAsString();

        assertEquals(inputJson, outputJson);
        assertThat(response.getStatus()).isEqualTo(HttpStatus.OK.value());
    }

    @Test
    public void testGetContractById() throws Exception {
        ContractDto contractDtoMock = ContractDto.builder().id(1L).build();

        when(contractService.findById(Mockito.anyLong())).thenReturn(contractDtoMock);

        RequestBuilder requestBuilder = MockMvcRequestBuilders
                .get(TEST_URL + "/1")
                .accept(MediaType.APPLICATION_JSON);

        MvcResult result = mockMvc.perform(requestBuilder).andReturn();
        String expectedJson = this.toJson(contractDtoMock);
        String outputJson = result.getResponse().getContentAsString();

        assertEquals(expectedJson, outputJson);
    }

    @Test
    public void testGetAllContracts() throws Exception {
        List<ContractDto> contractsDtoMock = new ArrayList<>();
        contractsDtoMock.add(ContractDto.builder().id(1L).build());
        contractsDtoMock.add(ContractDto.builder().id(2L).build());
        contractsDtoMock.add(ContractDto.builder().id(3L).build());

        when(contractService.findAll()).thenReturn(contractsDtoMock);

        RequestBuilder requestBuilder = MockMvcRequestBuilders
                .get(TEST_URL)
                .accept(MediaType.APPLICATION_JSON);

        MvcResult result = mockMvc.perform(requestBuilder).andReturn();
        String expectedJson = this.toJson(contractsDtoMock);
        String outputJson = result.getResponse().getContentAsString();

        assertEquals(expectedJson, outputJson);
    }

    @Test
    public void testDeleteContractById() throws Exception {
        ContractDto contractDtoMock = ContractDto.builder().id(1L).build();

        when(contractService.findById(Mockito.anyLong())).thenReturn(contractDtoMock);

        RequestBuilder requestBuilder = MockMvcRequestBuilders
                .delete(TEST_URL)
                .accept(MediaType.APPLICATION_JSON);

        MvcResult result = mockMvc.perform(requestBuilder).andReturn();
        MockHttpServletResponse response = result.getResponse();

        assertThat(response.getStatus()).isEqualTo(HttpStatus.OK.value());
    }

    private String toJson(Object object) throws JsonProcessingException {
        ObjectMapper objectMapper = new ObjectMapper();
        return objectMapper.writeValueAsString(object);
    }
}
